chore(frontend): extract replay-iframe-storage util to break heatmaps coupling
`heatmapsBrowserLogic` only needs one helper from `sessionRecordingPlayerLogic`
(`removeReplayIframeDataFromLocalStorage`), but pulling it in dragged the
entire ~33KB player logic file into the heatmaps scene's graph as a
transitive dependency. The player logic in turn pulls posthog-rrweb,
player settings, snapshot data logic, and so on.

Moves the three helpers (`ReplayIframeDatakeyPrefix`,
`removeFromLocalStorageWithPrefix`, `removeReplayIframeDataFromLocalStorage`)
into a tiny standalone `replay-iframe-storage.ts`. Heatmaps imports
from there directly. The player logic now imports the same helpers and
re-exports `removeReplayIframeDataFromLocalStorage` for any other consumer
that still expects it from the legacy path.

Also tightens `import { ReplayIframeData }` in the player logic to
`import type` so the heatmaps logic file is no longer pulled into the
player scene's graph.

This on its own does not move the shared mega-chunk because other
non-replay scenes still anchor session-recording code there. But it
removes a leak that would otherwise re-attach the dependency the moment
any of the other lazy-loading work lands.

Generated-By: PostHog Code
Task-Id: fff62b94-2da1-4084-8703-1bc35ece8875
